Updated input + fixed network error
The waitForNavigation call was hanging, I think due to a CORS error. Easy solution was just to comment it out.

I automated the date selection, as one will probably always want to check the upcoming few months.

I also included a prompt for the user to select which embassy to ping.